+++ ut_assert/inc/uttest.h
@@ -29,14 +29,41 @@
  * Exported Functions
  */
 
-/* Adds a new unit test to the test database. */
+/**
+ * \brief Adds a new unit test to the test database.
+ *
+ * Called by the user to register a new test case with the library.
+ *
+ * \param Test     Main test function to call.
+ * \param Setup    Setup function, called before the test function
+ * \param Teardown Cleanup function, called after the test function
+ * \param TestName Name of test for logging purposes
+ */
 void    UtTest_Add(void (*Test)(void), void (*Setup)(void), void (*Teardown)(void), const char *TestName);
 
-/*
- * Set up function for UT-Assert based test routines
- * This should call UtTest_Add() for each test set
+/**
+ * \brief Early initialization function
+ *
+ * Reset the global data to a safe state for initial start-up.
+ * This should be called before any other API.
  */
-void    UtTest_Setup(void);
+void UtTest_EarlyInit(void);
+
+
+/**
+ * \brief Execute all registered tests
+ *
+ * All test functions that were registered with UtTest_Add will be executed in order.
+ */
+void UtTest_Run(void);
+
+/*
+ * \brief Set up function for UT-Assert based test routines
+ *
+ * This function must be provided by the user to set up test cases.
+ * This should call UtTest_Add() for each test case.
+ */
+void UtTest_Setup(void);
 
 #endif
